requirements.txt
python-telegram-bot==20.7
PyMuPDF
tiktoken
faiss-cpu
openai
python-dotenv

